craigslister/requirements.txt
beautifulsoup4==4.5.1
py-dateutil==2.2
python-craigslist==1.0.3
requests==2.11.1
six==1.10.0
SQLAlchemy==1.0.15
slackclient
